Fix #142728
