+++ src/vs/workbench/contrib/files/browser/fileCommands.ts
@@ -17,7 +17,7 @@ import { IClipboardService } from 'vs/platform/clipboard/common/clipboardService
 import { toErrorMessage } from 'vs/base/common/errorMessage';
 import { IListService } from 'vs/platform/list/browser/listService';
 import { CommandsRegistry } from 'vs/platform/commands/common/commands';
-import { RawContextKey, IContextKey, IContextKeyService, ContextKeyExpr } from 'vs/platform/contextkey/common/contextkey';
+import { IContextKey, IContextKeyService, ContextKeyExpr } from 'vs/platform/contextkey/common/contextkey';
 import { IFileService } from 'vs/platform/files/common/files';
 import { KeybindingsRegistry, KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegistry';
 import { KeyMod, KeyCode, KeyChord } from 'vs/base/common/keyCodes';
@@ -47,49 +47,7 @@ import { hash } from 'vs/base/common/hash';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IPaneCompositePartService } from 'vs/workbench/services/panecomposite/browser/panecomposite';
 import { ViewContainerLocation } from 'vs/workbench/common/views';
-
-// Commands
-
-export const REVEAL_IN_EXPLORER_COMMAND_ID = 'revealInExplorer';
-export const REVERT_FILE_COMMAND_ID = 'workbench.action.files.revert';
-export const OPEN_TO_SIDE_COMMAND_ID = 'explorer.openToSide';
-export const OPEN_WITH_EXPLORER_COMMAND_ID = 'explorer.openWith';
-export const SELECT_FOR_COMPARE_COMMAND_ID = 'selectForCompare';
-
-export const COMPARE_SELECTED_COMMAND_ID = 'compareSelected';
-export const COMPARE_RESOURCE_COMMAND_ID = 'compareFiles';
-export const COMPARE_WITH_SAVED_COMMAND_ID = 'workbench.files.action.compareWithSaved';
-export const COPY_PATH_COMMAND_ID = 'copyFilePath';
-export const COPY_RELATIVE_PATH_COMMAND_ID = 'copyRelativeFilePath';
-
-export const SAVE_FILE_AS_COMMAND_ID = 'workbench.action.files.saveAs';
-export const SAVE_FILE_AS_LABEL = nls.localize('saveAs', "Save As...");
-export const SAVE_FILE_COMMAND_ID = 'workbench.action.files.save';
-export const SAVE_FILE_LABEL = nls.localize('save', "Save");
-export const SAVE_FILE_WITHOUT_FORMATTING_COMMAND_ID = 'workbench.action.files.saveWithoutFormatting';
-export const SAVE_FILE_WITHOUT_FORMATTING_LABEL = nls.localize('saveWithoutFormatting', "Save without Formatting");
-
-export const SAVE_ALL_COMMAND_ID = 'saveAll';
-export const SAVE_ALL_LABEL = nls.localize('saveAll', "Save All");
-
-export const SAVE_ALL_IN_GROUP_COMMAND_ID = 'workbench.files.action.saveAllInGroup';
-
-export const SAVE_FILES_COMMAND_ID = 'workbench.action.files.saveFiles';
-
-export const OpenEditorsGroupContext = new RawContextKey<boolean>('groupFocusedInOpenEditors', false);
-export const OpenEditorsDirtyEditorContext = new RawContextKey<boolean>('dirtyEditorFocusedInOpenEditors', false);
-export const OpenEditorsReadonlyEditorContext = new RawContextKey<boolean>('readonlyEditorFocusedInOpenEditors', false);
-export const ResourceSelectedForCompareContext = new RawContextKey<boolean>('resourceSelectedForCompare', false);
-
-export const REMOVE_ROOT_FOLDER_COMMAND_ID = 'removeRootFolder';
-export const REMOVE_ROOT_FOLDER_LABEL = nls.localize('removeFolderFromWorkspace', "Remove Folder from Workspace");
-
-export const PREVIOUS_COMPRESSED_FOLDER = 'previousCompressedFolder';
-export const NEXT_COMPRESSED_FOLDER = 'nextCompressedFolder';
-export const FIRST_COMPRESSED_FOLDER = 'firstCompressedFolder';
-export const LAST_COMPRESSED_FOLDER = 'lastCompressedFolder';
-export const NEW_UNTITLED_FILE_COMMAND_ID = 'workbench.action.files.newUntitledFile';
-export const NEW_UNTITLED_FILE_LABEL = nls.localize('newUntitledFile', "New Untitled File");
+import { OPEN_TO_SIDE_COMMAND_ID, COMPARE_WITH_SAVED_COMMAND_ID, SELECT_FOR_COMPARE_COMMAND_ID, ResourceSelectedForCompareContext, COMPARE_SELECTED_COMMAND_ID, COMPARE_RESOURCE_COMMAND_ID, COPY_PATH_COMMAND_ID, COPY_RELATIVE_PATH_COMMAND_ID, REVEAL_IN_EXPLORER_COMMAND_ID, OPEN_WITH_EXPLORER_COMMAND_ID, SAVE_FILE_COMMAND_ID, SAVE_FILE_WITHOUT_FORMATTING_COMMAND_ID, SAVE_FILE_AS_COMMAND_ID, SAVE_ALL_COMMAND_ID, SAVE_ALL_IN_GROUP_COMMAND_ID, SAVE_FILES_COMMAND_ID, REVERT_FILE_COMMAND_ID, REMOVE_ROOT_FOLDER_COMMAND_ID, PREVIOUS_COMPRESSED_FOLDER, NEXT_COMPRESSED_FOLDER, FIRST_COMPRESSED_FOLDER, LAST_COMPRESSED_FOLDER, NEW_UNTITLED_FILE_COMMAND_ID, NEW_UNTITLED_FILE_LABEL } from './explorerConstants';
 
 export const openWindowCommand = (accessor: ServicesAccessor, toOpen: IWindowOpenable[], options?: IOpenWindowOptions) => {
 	if (Array.isArray(toOpen)) {
